requirements.txt
asyncio==3.4.3
buttons==0.1.9
chat-exporter==1.3.0.6.5
discord.py @ git+https://github.com/Rapptz/discord.py@fc51736b34df3c37c06781ae948b18efb71616fa
DiscordUtils==1.2.8
discord-better-help==1.3.0
dnspython==1.16.0
humanfriendly==9.2
emojis==0.6.0
python-dotenv==0.17.1
eight-ball==3.0.3
motor==2.3.1
Paginator==0.5.1
python-dateutil==2.8.1
python-string-utils==1.0.0
psutil==5.8.0
bs4==0.0.1
